[refactor] 조회 api 분리
