+++ src/main/java/com/planpal/demo/service/schedule/ScheduleService.java
@@ -18,7 +18,6 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.ArrayList;
 import java.util.List;
 
 @Service
@@ -97,29 +96,6 @@ public class ScheduleService {
                 addedSchedule.setIsChecked(false);
             }
         }
-    }
-
-    /*
-    * 하나의 일정 조회
-    * */
-    public GetScheduleResponse getSchedule(Long scheduleId){
-        Schedule schedule=schedulesRepository.findById(scheduleId)
-                        .orElseThrow(() -> new ScheduleException(ErrorStatus.SCHEDULE_NOT_FOUND));
-        return ScheduleConverter.toGetScheduleResponse(schedule);
-    }
-
-    /*
-    * 간단 일정 전체 조회
-    * */
-    public GetAllScheduleListResponse getAllSimpleScheduls(){
-        List<Schedule> schedules = schedulesRepository.findAll();
-        List<GetSimpleScheduleResponse> simpleSchedules=new ArrayList<>();
-        for (Schedule schedule : schedules){
-            GetSimpleScheduleResponse simplesSchedule=ScheduleConverter.toSimpleSchedule(schedule);
-            simpleSchedules.add(simplesSchedule);
-        }
-
-        return ScheduleConverter.toSimpleScheduleList(simpleSchedules);
     }
 
     /*
