+++ src/main/java/com/planpal/demo/web/controller/schedule/ScheduleGetController.java
@@ -2,7 +2,7 @@ package com.planpal.demo.web.controller.schedule;
 
 import com.planpal.demo.apipayload.ApiResponse;
 import com.planpal.demo.apipayload.status.SuccessStatus;
-import com.planpal.demo.service.schedule.ScheduleService;
+import com.planpal.demo.service.schedule.ReadScheduleService;
 import com.planpal.demo.web.dto.schedule.GetAllScheduleListResponse;
 import com.planpal.demo.web.dto.schedule.GetScheduleResponse;
 import lombok.RequiredArgsConstructor;
@@ -13,17 +13,17 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @RequiredArgsConstructor
 public class ScheduleGetController {
-    private final ScheduleService scheduleService;
+    private final ReadScheduleService readScheduleService;
 
     @GetMapping("/schedules/{scheduleId}")
     public ApiResponse<GetScheduleResponse> getSchedule(@PathVariable Long scheduleId){
-        GetScheduleResponse response=scheduleService.getSchedule(scheduleId);
+        GetScheduleResponse response=readScheduleService.getSchedule(scheduleId);
         return ApiResponse.of(SuccessStatus._OK, response);
     }
 
     @GetMapping("/schedules")
     public ApiResponse<GetAllScheduleListResponse> getAllSchedules(){
-        GetAllScheduleListResponse response=scheduleService.getAllSimpleScheduls();
+        GetAllScheduleListResponse response=readScheduleService.getAllSimpleScheduls();
         return ApiResponse.of(SuccessStatus._OK, response);
     }
 }
